Fix for DIRSERVER-1475
git-svn-id: https://svn.apache.org/repos/asf/directory/apacheds/trunk@919009 13f79535-47bb-0310-9956-ffa450edef68

+++ core-api/src/main/java/org/apache/directory/server/core/CoreSession.java
@@ -699,7 +699,7 @@ public interface CoreSession
      * @throws Exception if there are failures while listing children
      */
     EntryFilteringCursor list( DN dn, AliasDerefMode aliasDerefMode, 
-        Set<AttributeTypeOptions> returningAttributes, int sizeLimit, int timeLimit ) throws Exception;
+        Set<AttributeTypeOptions> returningAttributes, long sizeLimit, int timeLimit ) throws Exception;
     
     
     /**
@@ -765,7 +765,7 @@ public interface CoreSession
      * @throws Exception if there are failures while listing children
      */
     EntryFilteringCursor search( DN dn, SearchScope scope, ExprNode filter, AliasDerefMode aliasDerefMode, 
-        Set<AttributeTypeOptions> returningAttributes, int sizeLimit, int timeLimit ) throws Exception;
+        Set<AttributeTypeOptions> returningAttributes, long sizeLimit, int timeLimit ) throws Exception;
 
 
     EntryFilteringCursor search( InternalSearchRequest searchRequest ) throws Exception;
