+++ core-mock/src/main/java/org/apache/directory/server/core/MockDirectoryService.java
@@ -275,13 +275,13 @@ public class MockDirectoryService implements DirectoryService
     }
 
 
-    public void setMaxSizeLimit( int maxSizeLimit )
+    public void setMaxSizeLimit( long maxSizeLimit )
     {
 
     }
 
 
-    public int getMaxSizeLimit()
+    public long getMaxSizeLimit()
     {
         return 0;
     }
