+++ core/src/test/java/org/apache/directory/server/core/authz/support/MaxImmSubFilterTest.java
@@ -702,13 +702,13 @@ public class MaxImmSubFilterTest
         }
 
 
-        public void setMaxSizeLimit( int maxSizeLimit )
+        public void setMaxSizeLimit( long maxSizeLimit )
         {
 
         }
 
 
-        public int getMaxSizeLimit()
+        public long getMaxSizeLimit()
         {
             return 0;
         }
